requirements.txt
# requirements.txt

numpy>=1.21.0
pytest>=6.0.0
biopython==1.81
pandas==2.1.3
scikit-learn==1.3.0